Update Sketch Preview Data fetching
This replaces the dependency `jszip-utils` with `axios`, as the former
is mainly used for some data fetching. All the cases it seems to handle
(e.g. some quirks for older versions of Internet Explorer) we do not
really care about and therefore we can just replace it with axios.

Also reworked the loading logic a little to be more readable.