Remove redundant variable conversions from CI config process
In CI config, we receive variables as key-value in YAML format.
Then, we convert them into array format and process. Sometimes,
we do this process over and over again.

Also, we add "public: true" to all variables from the YAML config.
"public" is already "true" for variable items, so we don't need it.

In this commit, we remove unnecessary conversions
and the public attribute.