Add GraphQL custom emoji schema, flagged
This change exposes GraphQL schema that was previously hidden by default
so has a Changelog.

The old `feature_flag` is deprecated and should be replaced with a
flag that toggles the value of fields and behaviour of mutations.

This also adds `alpha` as the flag is disabled by default.

https://gitlab.com/gitlab-org/gitlab/-/issues/369202

Changelog: added