Replace native date input field to `GlDatePicker`
In `Profile > SSH keys`, I replaced the native date field with a
`ExpiresAtField` Vue component (which uses `GlDatePicker`).
`ExpiresAtField` renders by default a date of 30 days in the future.
That was an security recommendation for access tokens (see
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/93128), but I am a
unsure if that would be beneficial in the SSH context.

Because `ExpiresAtField` doesn't horizontally align with the title field
I decided to move to the next row.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/271239

Changelog: changed