Report full size of Ci::Pipeline hierarchy trees
By using #all_objects, we only count the single upstream and all
downstreams of a given downstream pipeline. For example, a tree
consisting of 1 root pipeline that directly creates 1,000
downstream pipelines will be reported as 1,000 pipelines with a
hierarchy of size 2. Instead, we will now log 1,001 pipelines all
with a hierarchy size of 1,001.