Fix flaky SSH remote mirror spec
This failed in https://gitlab.com/gitlab-org/gitlab/-/jobs/2905098518,
likely needing the same fix in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/92178.

The test was failing intermittently because the test would input the
URL and submit the data faster than the JavaScript was able to update
the internal field. This caused a blank URL to be submitted on
occasion. To make this spec more reliable, we now ensure the hidden
URL field is updated with the right value before clicking the "Save
changes" button.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/337394