Exclude variables with undefined description as prefill
When backend sends config variables, it sends "description: null" when
a variable does not have a description. With this, frontend does not
display them. However, it is checked as "description !== null", so
when description is undefined, the variable can be shown as a prefill.

In this commit, we are changing this condition to "description != null"
to exclude undefined values.