Show deployment approval options for read access
Users with read access to deployments may want to see what comments have
been left on a deployment. The comment box and approve/reject buttons
are still visible, but the backend stops them from being able to approve
or reject a deployment that they do not have permissions for.

Changelog: changed