Set all attachments to Content-Type application/octet-stream
In GitLab 11.6 via
https://gitlab.com/gitlab-org/gitlab-workhorse/-/merge_requests/335,
Workhorse attempts to detect a blob's `Content-Type` by proxying the
download and examining the first 512 bytes. This was done to thwart a
security issue where certain types of files could trick the browser
into displaying an `inline` file and execute
(https://gitlab.com/gitlab-org/gitlab-foss/-/issues/36103).

However, the detection mechanism used relies on Go's
`http.DetectContentType([]byte)` function, which implements the
algorithm described in https://mimesniff.spec.whatwg.org/. This
detection mechanism only can detect a small number of types and can
cause files to be labeled with the wrong `Content-Type`.

For example, this Workhorse change caused Microsoft Word `.docx` files
to have a `Content-Type` of `application/zip` instead of
`application/vnd.openxmlformats-officedocument.wordprocessingml.document`. A
similar problem exists for other documents.

It's natural to assume that we have to fix the Content-Type detection,
but there's a simpler way that preserves both security and accuracy:
let the browser deal with it. For example, when a `.docx` file is
downloaded with this type, it seems that Firefox and Chrome use the
Content-Disposition, URL, and Content-Type as hints about what to do
with the file.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/26448

Changelog: fixed