Hide create project tile for some users
For users unable to create projects, the welcome state still includes a
tile to create new projects. The tile has the title `Create a project`
but it is not clickable and it is not grey-out (to show that the tile is
inactive).

There are two type of users unable to create projects:
* External users
* Regular users whose `Projects limit` setting has been set to zero

Here, we replace the create project tile by an alert.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/334121

Changelog: changed