Fetch test suite from report in parser
This changes Ci::Build#collect_test_reports!
to be consistent with other collect_*_reports!
methods.

It takes a TestReport as a parameter and
injects the test data from job artifacts
into the test report.

As it previously returned a TestSuite,
the callers of this method are changed
to get the relevant TestSuite from
the TestReport.