Remove extra space after dismissing free cap alert
This fixes an issue where some extra space would be left at the top of
the Usage Quotas page after dismissing the free cap alert.
To address this, we are removing some padding from the alert's container
and applying it to the alert itsefl as margin. To make sure the
container reacts to those changes, its overflow is set to auto.

Changelog: fixed
EE: true