Fixes the start and end IDs used for the migration
This migration iterates over system note metadata so we should use these
IDs instead of the IDs of the associated notes.

These are not always in sync because if other specs have run before this
that create notes without system note metadata, then the notes IDs would
be different from the system note metadata IDs.