Alter when to send releasedAt when release created
Previously, we started to send the releasedAt timestamp, defaulted to
the time the page was loaded, when creating a release. Unfortunately,
the API is quite simple in its comparison on whether or not a release is
historical, and any time difference would mean that the release was
Historical and no release evidence gathered.

Here, we save the original released at date upon page load (or upon
release fetch), and if the released at date isn't touched (at all, mind
you), then we do not send it.

This is kind of a quick fix for the issue, as if users alter the
released at timestamp to a different day and back to the current day, or
even to the current day at all, a new date is recorded, and we're back
to the issue. I think some refinement of the API might be necessary
here, but we can iterate on this.

Changelog: fixed