Swap access_level for role in access token entity
These new serializable group and project entities are going to be used
in conjunction with a Vue component. This frontend component consumes
`role` the `human_access` (string) version of the `access_level`
(integer).

I have asked if there was a JavaScript function to convert from the
integer version to the string version, and I was recommended to make the
change in the backend instead:
https://gitlab.slack.com/archives/C0GQHHPGW/p1661259205181069

The group and project entities are not yet in use.