Fix missing metadata for system notes
This fixes an issue where `SystemNoteMetadata` for a `Note` related to
tasks or issue type changes have not been saved as the model was
invalid.

This was visible for users on the activity stream of the issue as it
didn't display an icon. We're now using the `link` icon for relating
tasks to issues.

While adding a test for checking persistence, the `issue_type` also
failed and got fixed. However, for `issue_type` we have not yet
committed to an icon.

Changelog: fixed