Adjust example for running git-fsck repository check
The example command given for performing a git-fsck manually on the Gitaly
filesystem has the "git" command run as root. In the Omnibus installation,
which is the assumed context based on the example command given, the git
data is owned by user "git", and git will produce a "dubious ownership"
error. Performing the git-fsck as the appropriate user avoids this error,
and prevents any permissions-related issues down the road if the user
happens to swap "fsck" with "gc" or any other writing-command.