Stop raising error when actor not used for YamlProcessor feature flags
In development and test environments, we raise an error when
YamlProcessor feature flag is called without a defined actor.

However, in unit tests of entries, we have to add `with_actor` to
each of them.

So, in this commit, we are restricting this requirement to
a spec helper. And for now, we are only calling it for
create_pipeline_service* specs to cover all pipeline creation domain.