Remove s3_omit_multipart_urls feature flag
In GitLab 15.0,
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/85306 introduced
the `s3_omit_multipart_urls` feature flag to skip the generation of
pre-signed multipart URLs if the Workhorse S3 client were used. This
fixed an bug where multipart URLs could be left in an incomplete
state.

This has been the default since 15.0, and this disabling of multipart
URLs does not seem to have caused an issue, so let's drop it.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/359718

Changelog: changed