Clear protected branch cache upon creation in test
The introduction of the protected branch caching via the
`hash_based_cache_for_protected_branches` feature flag could cause
tests to fail since the protected branch cache might have the wrong
values.

Normally the cache is refreshed since protected branches are created
with `ProtectedBranches::CreateService`, but in tests FactoryBot
creates them without the service. To ensure the cache is in a good
state in tests, mimic this cache clearing after the creation of a
protected branch in the factory.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/371944