Remove unused project reference
IterationReferenceFilter had a code path that attempts to
reference an iteration's project. The code path was
never used and is removed in this commit.

Although iterations may be presented within a project context,
they, as domain objects, cannot/do not belong to projects
(for historical context, project-level iterations have been
implemented but never rolled out as a feature.)

EE: true