Reset access token form selectively
In the personal access token page, the default date is removed after
successfully creation of a token. That's because we reset the whole
form.

In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/93128, we
introduced a default date for access tokens because we want to
discourage users from creating never expiring tokens.

With this change the expiration date is preserved after successful
creation of a token. Ideally, the datepicker should be reseted to the
default date, but this is difficult at the moment, because these are two
independent Vue components. So instead, as a compromise, we preserve the
date, if one is used, after successful creation of a token.

In the near future, we are planning to convert the whole form into a Vue
component. At that point, it would be easier to reset the datepicker to
the default date.

Changelog: changed