Add SBoM insertion tasks
Add bulk insertion tasks which use
`Gitlab::Ingestion::BulkInsertableTask` in order to perform
a series of bulk upserts. Each task takes a mapping of relations (called
an OccurrenceMap) and turns it into attributes for each model. The
BulkInsertableTask takes the array of attributes and uses it to perform
a bulk upsert. Each task is idempotent, using `unique_by` in order to
distinguish between objects.