Fix subgroup support for approval rules
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/371106

**Problem**

We had a problem with inherited permissions for subgroups. Users
couldn't see appovers for groups if they don't have a direct membership
in them. It was fixed in scope of
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/91598.

However, there is a similar invalid check that filters subgroups with
inherited permissions out.

**Solution**

* `public_or_visible_to_user` method doesn't correctly support subgroup
permissions. Instead we can verify if user has an access to the group.
* Add `accessible_to_user` scope that correctly supports subgroups

Changelog: fixed
EE: true