Refactor CI Variables entry
Entry::Variables is getting complicated when we add another feature
to it. For now, each variable accept either a single alphanumeric value
or a hash including value-description. We plan to add the "expand"
attribute to it. Also, we plan to support an array of values.

In this commit, we refactored it to prepare those changes.

- We create Entry::Variable and use it in Entry::Variables
- We currently have SimpleVariable and ComplexVariable
- In order to allow other data attributes, we change
from "use_value_data" to "allowed_value_data".