Support parsing table of contents in gfm parser
Implement a plugin to parse GitLab Flavored Markdown
table of contents in the client-side GLFM markdown
parser.

This MR also provides two renderers for table of contents.
One of the renderers converts the table of contents into
a section element. The other creates a tableofcontents
element suitable for usage in the Content Editor