Fix page javascript after a deploy token got created
This create_deploy_token "page" is only rendered as response to the
"New deploy token" form submit.
It shows the secret token to the user one time, but is otherwise
identical with the Settings/Repository page.
This is why we now just import the other page's JavaScript here.

Changelog: fixed