Refactor Groups Rest API to use workhorse for avatar upload
This change set adds specs for the Groups Rest API to test the
avatar upload.

More importantly though it changes the avatar upload to be handled by workhorse.
This is mostly as a preparation for allow to remove group avatars via Rest API,
similar to what will be possible for Projects with
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/92604
and which is already possible for Topics with
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/73969

A follow up MR will implement the avatar removal and the same for the Users API.

Changelog: other