Add specs to detect extra audit events
We have an epic, ["Don't create audit events when there is no change"][0]

In short, we want to make sure we do NOT emit audit events for API
update calls that end up not resulting in an actual data change

As part of this epic, `EE::Projects::UpdateService` audit events where
identified as a potential place where the behavior described in the
epic was occurring.

While working on on that [issue][1], I added a spec to check for the
duplicate audit event behavior. It turned out that this audit event
source already does the correct thing.

As such, this commit just adds the aforementioned specs to prevent
possible future regressions

[0]:https://gitlab.com/groups/gitlab-org/-/epics/8547
[1]:https://gitlab.com/gitlab-org/gitlab/-/issues/370703

mr: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/96265
closes: https://gitlab.com/gitlab-org/gitlab/-/issues/370703