Clean up orphaned invited members
Prior to GitLab 14.1.2 (?), it was possible to get entries in the
`members` table that had the following properties:

1. `user_id` is null
2. `invite_token` is null
3. `invite_accepted_at` is non-null

These entries could cause 500 errors when viewing the member list via
the UI or API since `user` was `nil`.

On GitLab.com, there are 1,336 entries like this, with the newest one
created on 2021-08-31.

It's not entirely clear how this happened, but no new orphaned entries
have been created since then.

This commit cleans up rows that meet the three criteria above. We
have to use a temporary index to ensure the deletion is fast.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/371804

Changelog: fixed