Move test repo to a directory not gitignored
Before this change, the test repo symlink is under a directory
that's in `workhorse/.gitignore`. As a result, when gitlab-foss
pulls changes from gitlab project, it doesn't include the symlink.
This causes workhorse tests in gitlab-foss project to fail.

To fix it, the test repo is moved to a directory that is not in
`.gitignore`.

This also fixes the issue when `make clean` is called and it
deletes the symlink as well.