Speed up Projects::JobsController spec
Creating a project is very expensive on Gitaly calls.
Before the change, the total test time is 15m09s,
49% of which is spent on GRPC::Core::Call#run_batch.

This changes the test such that the project and users
are created once in the entire tests.

Each test can switch between the user roles
in order to test the permission level.

As a result, the total test time is now 2m29s and
GRPC::Core::Call#run_batch only takes 6.48% of it.