Implement workhorse per-key keywatcher channels
This adds feature flagged new behavior to the Redis keywatcher in
Workhorse. The keywatcher is (only) used for CI long polling.

The old behavior is that Workhorse subscribes to a single Redis pubsub
channel `workhorse:notifications` which acts as a notification
firehose. The new behavior is that for each key that Workhorse starts
watching, it subscribes to `workhorse:notifications:$key`. The
behavior is a toggled at startup via the environment variable
GITLAB_WORKHORSE_REDIS_SUBSCRIBE_MANY=1.

Workhorse will either listen to the old firehose channel or to the new
per-key channels but not both.