Improve bulk issue creation on epics
When creating issues on epics, it's common to create the issue in the
same project. With this change we no longer re-mount the issue creation
form and only reset the `title` of the issue.

We also add event listeners on `enter` so it's easy to create issues
with a keyboard.

Changelog: changed
EE: true