Geo: Realign concurrency limits
It's a likely unnoticed bug. Prior to this change, across many GitLab
versions, as more and more Models came to be replicated by the Geo
Self-Service Framework, the peak number of concurrent sync jobs appeared
to shrink below the specified limits.

A risk is that the readjustment will generally increase the peak number
of concurrent sync jobs.

Changelog: fixed
EE: true