Add logic to persist / retrieve baseURL
When we use the multiversion select to choose which
version of GitLab we want to use (.com or self-hosted),
we need to reload the page in order to update the
Content Security Policy from the backend to include
the new URL of the self-hosted instance.

Thus we need to store the baseURL to localStorage on
initial save, reload the page then retrieve it.