workhorse: Improve perf for LSIF hover entries with markdown
lsif-go and other tools commonly output a 'markdown' hover entry, rather than
specifying a language alongside a fragment of code. Unfortunately, the lexer
library we use to process code fragments enters a very slow code path if no
language, or an unknown language, is specified.

The reason for the slow path is because if no language is immediately matched
against an index, it starts to use filepath matching, using the language name
as either a filename or as a filename extension. It's very unlikely that the
language provided will be a filename, and if it is a common extension, this
will most likely already be registered and indexed as an alias.

This change bails early if no language is provided, and guards against the
slow filepath matches. A benchmark has been added to help detect such a
scenario.