Upgrade lookbook gem to v1.0
This version contains my fix that disables CSP in development mode.
So we can remove the class_eval we had before, and with that,
the entire initializer file.

Pages are not an experimental feature anymore, and the only
remaining config can go into config/development.rb together
with the view_component config.