Remove TestKeyChangesBubblesUpError test
This test was added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/86793 when we
thought we wanted to detect redis.ErrNil higher up the callstack.
However, during review of that MR we changed the implementation so
that redis.ErrNil is handled locally. We no longer expect redis.ErrNil
to bubble up the call stack so we don't need this test anymore.