Improve error message while validating config/database.yml
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/89696
we introduced additional handling of an error when doing an
upsert on a read-only database.

When a user encounters this scenario, they are presented with
a message stating "ERROR: cannot execute INSERT in a read-only
transaction". The verbiage is confusing to users because it
implies it is ERROR severity, but it should be a WARNING since
this does not break anything or halt any processes.

Changelog: other