Fix migration type
With https://gitlab.com/gitlab-org/gitlab/-/merge_requests/95881 a new
background migration was scheduled in a regular Rails migration. This
results in teh workers trying to execute the migration job while the job
class is not yet available, as teh deployment is not finished.

This move the scheduling migration under `db/post_migrate` to fif the
problem.

See
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/95881#note_1087932448.

Changelog: changed