Make `GITALY_SERVER_VERSION` approvable by maintainers
Currently `GITALY_SERVER_VERSION` can only be approved by
@project_278964_bot6 which cases an issue if the Gitaly version
update has an issue with Rails compatibility.

This makes it impossible to fix, as only the release-bot user
can approve, and non of maintainers or delivery members.

This fixes that by allowing the version update by maintainers
and delivery members which was the pattern used before.