Allow disabling the sidekiq-cron poller
In order to enqueue sidekiq-cron scheduled jobs, a poller runs in
Sidekiq at a configurable interval (30 seconds by default) and retrieves
all the cron jobs configured in redis. In a setup with multiple Sidekiq
instances this can result in a significant amount of load in redis (see
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1932). We
introduce a setting that allows specifying a cron polling interval, and
disables polling at all if the interval is 0.

In the future this will also enable having multiple Sidekiq clusters
pointing at different redis instances, by disabling the sidekiq-cron
poller in all but one of those clusters (see
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1439).