Introduce backend updates for toggle of diff preview
Add a column to both the project and the namespace settings which will
control whether a comment on a line in a merge request will show
the code section that goes with the comment.

Add necessary parts of "back end" code to utilize this column.
Front end will follow after this merge.

Changelog: added