Implement partition management logic for `security_findings` table
As the `security_findings` table is now partitioned, with this change
we are implementing the logic to create missing partitions and detach
redundant ones.

Basically, we create a new partition when the data size of existing
partition exceeds 10 Gigabytes and we detach a partition when all the
related `security_scans` for that partition becomes purged.