Add step to update instance_url for installation
Needed for updating the installation of self-managed
instances. Once the request is successful, we need to
reload the page so the backend can add the correct
Content Security Policy which includes the new
self-hosted URL.

Also properly reset the baseURL when GitLab
version is changed from GitLab.com to self-hosted
(and vice-versa).