Add OAuth flow for self-hosted
Step 1: Open a new window with the URL of the OAuth
application on the self-hosted instance. This requires
generating an authorization URL that includes a challenge.
Step 2: User signs in and authorizes from the Oauth app.
Step 3: User is redirected back to the main app (usually
GitLab.com) and the window closes automatically.
Step 4: The OAuth window returns a code which is used
to fetch an access token.