Remove GITLAB_LEGACY_BACKGROUND_UPLOADS
When we removed background upload for object storage in GitLab 15.0 we
left behind an escape hatch in case anyone got in trouble because of
this change after upgrading. It has been 4 months and we think it is
time to remove the escape hatch.

This commit removes support for the GITLAB_LEGACY_BACKGROUND_UPLOADS
environment variable.

Also see epic https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/734.

Changelog: other