Refactor Users Rest API to use workhorse for avatar upload
This change set adds specs for the Users Rest API to test the
avatar upload.
More importantly though it changes the avatar upload to be handled by workhorse.
This is mostly as a preparation for allow to remove ers avatars via Rest API,
similar to what will be possible for Projects with
gitlab-org/gitlab!92604
, Groups with https://gitlab.com/gitlab-org/gitlab/-/merge_requests/96381
and which is already possible for Topics with
gitlab-org/gitlab!73969

A follow up MR will implement the avatar removal.

The main trigger for this change is to fully support user avatar management via IaC using the GitLab Terraform provider.