Lower the default Rails.cache ttl to 8 hours
8 hours is the configuration we settled on for GitLab.com.

This helps keeping Redis' memory usage lower, which in turn avoids it
having to evict a lot of keys when it reaches maxmemory. In
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1601 we
have determined that these evictions cause a reduced throughput on the
Redis-side. Which in turn makes the clients wait, causing requests to
be slower during one of these eviction bursts.

In https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1854
we've gradually lowered the TTL to where these eviction bursts were no
longer a problem, while not affecting the cache hit ratio or overall
request latency.

The traffic patterns of self-managed instances might be very
different, but 8 hours (a working day) is likely a good default for
most cases.

Changelog: changed