Simplify primary navigation by removing infrequently used options
Some menu options are not used often and merely serve to create visual
noise. Many of these are redundant or lead to the same pages with a different
tab selected. We are aiming to simplify navigation and reduce confusion by
culling unnecessary options.

This change is governed by a default-enabled feature flag:
remove_extra_primary_submenu_options

Changelog: changed