Remove and replace old images; rewrite page
Not a 100% rewrite of the page, but when I looked at the images, I
realized we had a really old page on our hands. When I tried to
recreate the steps to do a cherry-pick, I couldn't follow what was
on the page, so I ended up rewriting the steps along with reshooting
the screencaps.