Fix Workhorse Redis Sentinel role checks
Workhorse was configured to check the Redis replication role of the
Redis server it connects to before every Redis command. This is too
often. The Redis documentation recommends to check the role just once,
immediately after establishing the connection.
https://redis.io/docs/reference/sentinel-clients/

This commit moves the role check from the TestOnBorrow callback to the
Dial callback.

Changelog: performance