Use distinct location fingerprints for manually created vulnerabilities
Manually created vulnerabilities all share a common, hard-coded location
fingerprint. This means that when two vulnerabilities are created with
the same primary identifier, they are considered duplicates because they
have the same UUID. This change generates a unique location fingerprint
for each manually created vulnerability, so that they are not considered
equivalent and different manual vulnerabilities may be created with the
same identifier.

Changelog: fixed
EE: true