Unlock CI pipeline artifacts when pipeline is unlocked
Currently the SQL query for finding expired CI pipeline artifacts (not
to be confused with CI job artifacts) is too expensive and is timing
out. In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/97194,
we added a `locked` column to help. When a pipeline is unlocked, it
will also unlock the associated pipeline artifacts. With this
denormalization, querying for expired and unlocked artifacts is fast.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/372500

Changelog: changed