Fix feature flag link in related feature flags
The link to the feature flag in the related feature flags block in
issues pointed to the detail route, but there is only an edit route.

Using the edit route is okay, as if a user can admin related feature
flags, they can also edit feature flags, and this block is conditional
on the ability to admin the block.

Changelog: fixed
EE: true