Detect concurrent updates to project statistics
Add a database record lock using exclusive lease on
project statistics updates, using `#try_with_record_lock`.

This method attempts to obtain a lock before making
updates on the database row. If the lock is not obtained,
it would log a warning to indicate that there has been
a concurrent update attempt. It would then fallback
to updating the row as usual.