Add internal column to notes
This adds the `internal` column to the `notes` table as we renamed
`confidential` to `internal`.
Both columns will exist until we migrated the data and can drop
`confidential`.

Compared to the `confidential` column, it's using
`DEFAULT false` and `NOT NULL` as it's easier to deal with only boolean
values. The `confidential` column didn't use a default values as it was
created with a PG version <11 where adding a default value was not
cheap.

Changelog: other