Resetting form doesn't affect submit button
Fix regression from
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/96379.

After a successful creation of a personal access token the text of the
submit button was removed.

We have change the logic to be more selective. In addition, we don't add
`checked` attributes to text input fields and, the other way around,
adding `value` attributes to checkbox input fields

Changelog: fixed