Add array notation to GitHub Import representation objects
Representation objects allow their properties to be accessed using dot
notation like Sawyer Resources that are used by Octokit, which allow
methods in GitHub Import to receive as parameter a representation object
or a Sawyer resource object.

Because Sawyer Resource objects will be converted to a hash in future
changes, this change makes representation objects to support the array
notation. So methods in GitHub Import will be able to receive
representation objects or hashes as parameters.