Remove :seed_helper from Git::Commit spec
SeedHelper uses git to directly manipulate repositories within gitaly.
Soon we will be enabling praefect by default and when this happens we
cannot modify repositories without praefects knowledge.

This is a best effort translation from seed helper to rpsec factories.
Most if not all of these specs already exist on the gitaly side. Since
the factory repository is slightly different it's not always possible to
exactly replicate the tests here.