Fix create pipeline service spec
We swapped the order of this file around because it was failing when run
on its own. This was due to the protected branch examples in the middle
of the file: they were persisting data in the Redis cache that then
continued to future examples, and led to stale data (reporting that the
project had some protected branches when there were none).

We can fix this by putting the specs back where they were and clearing
the cache between each spec.