Add more options for feature flag changelogs in docs
Feature flags are not only used for feature rollouts, they are also
used for rolling out bug fixes and maintenance works. However, in the
feature flag changelog documentation, only "added" and "changed" are
allowed for enabling the feature flag. In this commit, we are adding
more options to the documentation.