Use a stacked layout for runners list
This MR enables the more compact layout for runners in a list.

The goal of this additional redesign is to improve consistency
throughout our table/list views and increase the flexibility of adding
more data to the table, as well as improving the readability of that
data.

This aligns the runner list to GitLab's new list design standards.

Changelog: changed