Cop/StaticTranslationDefinition: Check multiline strings
To support multiline strings we need to allow dynamic strings as well.
Both are `dstr`.

With this, we check for:

  _('long string ' \
    'ends here')

In Ruby AST it's a `dstr` but later it's just a plain string 'long
string ends here' and thus normally used by gettext.

Now, we also check for:

  _("something with #{interpolation}")

This is because it's also `dstr` from AST point of view.
Interpolated strings are not recognized by gettext. This fact we have to
catch in a separate cop.