Refactor environments polling mechanism
Due to apollo triggering a strange polling loop when combining a
reactive polling interval and changing query variables, this is being
refactored to provide a static polling interval of 3 seconds.

Then, the polling interval variable sets up a watcher that changes the
polling rate as the interval changes. It should only be triggered once
when the initial polling interval is loaded, but this does seem to avoid
the looping scenario somehow.

resetPolling is renamed to refetchEnvironments, as it calls refetch when
an environment is changed (stopped, approved, etc.) to speed up the UI
updates when the user is the one interacting with the environment.